backend/requirements.txt
python==3.12.*
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
langchain
langchain_community
httpx
grpcio
grpcio-tools
yandexcloud
pandas
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
